Move search subscription types to /models
