+++ src/client.spec.ts
@@ -22,8 +22,6 @@ import {
 	ScheduledQueriesFilter,
 	ScheduledScriptsFilter,
 	SearchDownloadFormat,
-	SearchFilter,
-	SearchSubscription,
 	SystemStatusCategory,
 	SystemStatusMessageReceived,
 	SystemStatusMessageSent,
@@ -73,7 +71,9 @@ import {
 	Script,
 	Search,
 	Search2,
+	SearchFilter,
 	SearchModule,
+	SearchSubscription,
 	TargetedNotificationTargetType,
 	Template,
 	UpdatableActionable,
