+++ src/functions/searches/subscribe-to-one-search/subscribe-to-one-search.ts
@@ -10,8 +10,8 @@ import { decode as base64Decode } from 'base-64';
 import { isArray, isBoolean, isEqual, isNull, isUndefined } from 'lodash';
 import { BehaviorSubject, combineLatest, Observable } from 'rxjs';
 import { distinctUntilChanged, filter, first, map } from 'rxjs/operators';
-import { Query } from '../../../models';
-import { ID, Percentage, toNumericID } from '../../../value-objects';
+import { Query, SearchEntries, SearchFilter, SearchStats, SearchSubscription } from '../../../models';
+import { Percentage, toNumericID } from '../../../value-objects';
 import { APIFunctionMakerOptions, promiseProgrammatically } from '../../utils';
 import {
 	isRawTableEntries,
@@ -232,58 +232,3 @@ export const makeSubscribeToOneSearch = (makerOptions: APIFunctionMakerOptions) 
 		};
 	};
 };
-
-export interface SearchFilter {
-	limit: number;
-	start: Date;
-	end: Date;
-}
-
-export interface SearchStats {
-	id: ID;
-	userID: ID;
-
-	entries: number;
-	duration: string;
-	start: Date;
-	end: Date;
-
-	pipeline: Array<{
-		module: string;
-		arguments: string;
-		duration: number;
-		input: {
-			bytes: number;
-			entries: number;
-		};
-		output: {
-			bytes: number;
-			entries: number;
-		};
-	}>;
-
-	storeSize: number;
-	processed: {
-		entries: number;
-		bytes: number;
-	};
-}
-
-// TODO: Add render module to entries observable
-export interface SearchEntries {
-	start: Date;
-	end: Date;
-
-	names: Array<string>;
-	data: Array<{
-		timestamp: Date;
-		values: Array<string | number | null>;
-	}>;
-}
-
-export interface SearchSubscription {
-	progress$: Observable<Percentage>;
-	entries$: Observable<SearchEntries>;
-	stats$: Observable<SearchStats>;
-	setFilter: (filter: SearchFilter) => void;
-}
