+++ src/models/search/index.ts
@@ -11,6 +11,10 @@ export * from './is-valid-search';
 export * from './raw-search';
 export * from './raw-search2';
 export * from './search';
+export * from './search-entries';
+export * from './search-filter';
+export * from './search-stats';
+export * from './search-subscription';
 export * from './search2';
 export * from './to-search';
 export * from './to-search2';
